+++ eq-author-api/src/businessLogic/createExpression.js
@@ -1,11 +1,9 @@
 const uuid = require("uuid");
 const createLeftSide = require("./createLeftSide");
-const createRightSide = require("./createRightSide");
 
 module.exports = input => ({
   id: uuid.v4(),
   condition: "Equal",
   left: createLeftSide(),
-  right: createRightSide(),
   ...input,
 });
