--- eq-author-api/src/businessLogic/createRightSide.js
@@ -1,6 +1,0 @@
-module.exports = input => ({
-  type: "SelectedOptions",
-  customValue: "",
-  optionIds: [],
-  ...input,
-});
