+++ eq-author-api/src/businessLogic/index.js
@@ -4,7 +4,6 @@ const createDestination = require("./createDestination");
 const createExpressionGroup = require("./createExpressionGroup");
 const createExpression = require("./createExpression");
 const createLeftSide = require("./createLeftSide");
-const createRightSide = require("./createRightSide");
 const getNextDestination = require("./getNextDestination");
 
 module.exports = {
@@ -14,6 +13,5 @@ module.exports = {
   createExpressionGroup,
   createExpression,
   createLeftSide,
-  createRightSide,
   getNextDestination,
 };
